google_cloud_functions/gather_metrics/requirements.txt
google-cloud-firestore==2.3.0
google-cloud-storage
